ssd1306_70_40 working! also I am bad at preprocessor directives lol
