+++ src/pico/pico-screens/screens/shared.h
@@ -13,7 +13,7 @@ extern "C" {
 #define DOOM_HEIGHT SCREENHEIGHT
 
 // here be some chicanery to avoid floats
-#define DOWNSAMPLING_FACTOR_OUT_OF_100 148
+#define DOWNSAMPLING_FACTOR_OUT_OF_100 500
 #define DOWNSAMPLED_WIDTH ((SCREENWIDTH * 100 / DOWNSAMPLING_FACTOR_OUT_OF_100))
 #define DOWNSAMPLED_HEIGHT ((SCREENHEIGHT * 100 / DOWNSAMPLING_FACTOR_OUT_OF_100))
