+++ src/pico/pico-screens/screens/ssd1306_70_40_i2c.c
@@ -15,100 +15,13 @@
 // +3 to center on 70 pixel width screen
 #define START_X (28 + 3)
 
-static ssd1306_t disp;
+static uint8_t buf[SSD1306_BUF_LEN];
 
-/**
- * Copyright (c) 2021 Raspberry Pi (Trading) Ltd.
- *
- * SPDX-License-Identifier: BSD-3-Clause
- */
-
-#include <stdio.h>
-#include <string.h>
-#include <stdlib.h>
-#include <ctype.h>
-#include "pico/stdlib.h"
-#include "pico/binary_info.h"
-#include "hardware/i2c.h"
-
-/* Example code to talk to an SSD1306-based OLED display
-
-   The SSD1306 is an OLED/PLED driver chip, capable of driving displays up to
-   128x64 pixels.
-
-   NOTE: Ensure the device is capable of being driven at 3.3v NOT 5v. The Pico
-   GPIO (and therefore I2C) cannot be used at 5v.
-
-   You will need to use a level shifter on the I2C lines if you want to run the
-   board at 5v.
-
-   Connections on Raspberry Pi Pico board, other boards may vary.
-
-   GPIO PICO_DEFAULT_I2C_SDA_PIN (on Pico this is GP4 (pin 6)) -> SDA on display
-   board
-   GPIO PICO_DEFAULT_I2C_SCL_PIN (on Pico this is GP5 (pin 7)) -> SCL on
-   display board
-   3.3v (pin 36) -> VCC on display board
-   GND (pin 38)  -> GND on display board
-*/
-
-// Define the size of the display we have attached. This can vary, make sure you
-// have the right size defined or the output will look rather odd!
-// Code has been tested on 128x32 and 128x64 OLED displays
-#define SSD1306_HEIGHT              64
-#define SSD1306_WIDTH               128
-
-#define SSD1306_I2C_ADDR            _u(0x3C)
-
-// 400 is usual, but often these can be overclocked to improve display response.
-// Tested at 1000 on both 32 and 84 pixel height devices and it worked.
-#define SSD1306_I2C_CLK             400
-//#define SSD1306_I2C_CLK             1000
-
-
-// commands (see datasheet)
-#define SSD1306_SET_MEM_MODE        _u(0x20)
-#define SSD1306_SET_COL_ADDR        _u(0x21)
-#define SSD1306_SET_PAGE_ADDR       _u(0x22)
-#define SSD1306_SET_HORIZ_SCROLL    _u(0x26)
-#define SSD1306_SET_SCROLL          _u(0x2E)
-
-#define SSD1306_SET_DISP_START_LINE _u(0x40)
-
-#define SSD1306_SET_CONTRAST        _u(0x81)
-#define SSD1306_SET_CHARGE_PUMP     _u(0x8D)
-
-#define SSD1306_SET_SEG_REMAP       _u(0xA0)
-#define SSD1306_SET_ENTIRE_ON       _u(0xA4)
-#define SSD1306_SET_ALL_ON          _u(0xA5)
-#define SSD1306_SET_NORM_DISP       _u(0xA6)
-#define SSD1306_SET_INV_DISP        _u(0xA7)
-#define SSD1306_SET_MUX_RATIO       _u(0xA8)
-#define SSD1306_SET_DISP            _u(0xAE)
-#define SSD1306_SET_COM_OUT_DIR     _u(0xC0)
-#define SSD1306_SET_COM_OUT_DIR_FLIP _u(0xC0)
-
-#define SSD1306_SET_DISP_OFFSET     _u(0xD3)
-#define SSD1306_SET_DISP_CLK_DIV    _u(0xD5)
-#define SSD1306_SET_PRECHARGE       _u(0xD9)
-#define SSD1306_SET_COM_PIN_CFG     _u(0xDA)
-#define SSD1306_SET_VCOM_DESEL      _u(0xDB)
-
-#define SSD1306_PAGE_HEIGHT         _u(8)
-#define SSD1306_NUM_PAGES           (SSD1306_HEIGHT / SSD1306_PAGE_HEIGHT)
-#define SSD1306_BUF_LEN             (SSD1306_NUM_PAGES * SSD1306_WIDTH)
-
-#define SSD1306_WRITE_MODE         _u(0xFE)
-#define SSD1306_READ_MODE          _u(0xFF)
-
-
-struct render_area {
-    uint8_t start_col;
-    uint8_t end_col;
-    uint8_t start_page;
-    uint8_t end_page;
-
-    int buflen;
+static struct render_area frame_area = {
+    start_col: 0,
+    end_col : SSD1306_WIDTH - 1,
+    start_page : 0,
+    end_page : SSD1306_NUM_PAGES - 1
 };
 
 void calc_render_area_buflen(struct render_area *area) {
@@ -116,135 +29,15 @@ void calc_render_area_buflen(struct render_area *area) {
     area->buflen = (area->end_col - area->start_col + 1) * (area->end_page - area->start_page + 1);
 }
 
-void SSD1306_send_cmd(uint8_t cmd) {
-    // I2C write process expects a control byte followed by data
-    // this "data" can be a command or data to follow up a command
-    // Co = 1, D/C = 0 => the driver expects a command
-    uint8_t buf[2] = {0x80, cmd};
-    i2c_write_blocking(i2c1, (SSD1306_I2C_ADDR & SSD1306_WRITE_MODE), buf, 2, false);
+// TODO move this to the screen library
+void ssd1306_70_40_i2c_clearScreen(void) {
+    memset(buf, 0, SSD1306_BUF_LEN);
 }
-
-void SSD1306_send_cmd_list(uint8_t *buf, int num) {
-    for (int i=0;i<num;i++)
-        SSD1306_send_cmd(buf[i]);
-}
-
-void SSD1306_send_buf(uint8_t buf[], int buflen) {
-    // in horizontal addressing mode, the column address pointer auto-increments
-    // and then wraps around to the next page, so we can send the entire frame
-    // buffer in one gooooooo!
-
-    // copy our frame buffer into a new buffer because we need to add the control byte
-    // to the beginning
-
-    uint8_t *temp_buf = malloc(buflen + 1);
-
-    temp_buf[0] = 0x40;
-    memcpy(temp_buf+1, buf, buflen);
-
-    i2c_write_blocking(i2c1, (SSD1306_I2C_ADDR & SSD1306_WRITE_MODE), temp_buf, buflen + 1, false);
-
-    free(temp_buf);
-}
-
-void SSD1306_init() {
-    // Some of these commands are not strictly necessary as the reset
-    // process defaults to some of these but they are shown here
-    // to demonstrate what the initialization sequence looks like
-    // Some configuration values are recommended by the board manufacturer
-
-    uint8_t cmds[] = {
-        SSD1306_SET_DISP,               // set display off
-        /* memory mapping */
-        SSD1306_SET_MEM_MODE,           // set memory address mode 0 = horizontal, 1 = vertical, 2 = page
-        0x00,                           // horizontal addressing mode
-        /* resolution and layout */
-        SSD1306_SET_DISP_START_LINE,    // set display start line to 0
-        SSD1306_SET_SEG_REMAP | 0x01,   // set segment re-map, column address 127 is mapped to SEG0
-        SSD1306_SET_MUX_RATIO,          // set multiplex ratio
-        SSD1306_HEIGHT - 1,             // Display height - 1
-        SSD1306_SET_COM_OUT_DIR | 0x08, // set COM (common) output scan direction. Scan from bottom up, COM[N-1] to COM0
-        SSD1306_SET_DISP_OFFSET,        // set display offset
-        0x00,                           // no offset
-        SSD1306_SET_COM_PIN_CFG,        // set COM (common) pins hardware configuration. Board specific magic number. 
-                                        // 0x02 Works for 128x32, 0x12 Possibly works for 128x64. Other options 0x22, 0x32
-#if ((SSD1306_WIDTH == 128) && (SSD1306_HEIGHT == 32))
-        0x02,                           
-#elif ((SSD1306_WIDTH == 128) && (SSD1306_HEIGHT == 64))
-        0x12,
-#else
-        0x02,
-#endif
-        /* timing and driving scheme */
-        SSD1306_SET_DISP_CLK_DIV,       // set display clock divide ratio
-        0x80,                           // div ratio of 1, standard freq
-        SSD1306_SET_PRECHARGE,          // set pre-charge period
-        0xF1,                           // Vcc internally generated on our board
-        SSD1306_SET_VCOM_DESEL,         // set VCOMH deselect level
-        0x30,                           // 0.83xVcc
-        /* display */
-        SSD1306_SET_CONTRAST,           // set contrast control
-        0xFF,
-        SSD1306_SET_ENTIRE_ON,          // set entire display on to follow RAM content
-        SSD1306_SET_NORM_DISP,           // set normal (not inverted) display
-        SSD1306_SET_CHARGE_PUMP,        // set charge pump
-        0x14,                           // Vcc internally generated on our board
-        SSD1306_SET_SCROLL | 0x00,      // deactivate horizontal scrolling if set. This is necessary as memory writes will corrupt if scrolling was enabled
-        SSD1306_SET_DISP | 0x01, // turn display on
-    };
-
-    SSD1306_send_cmd_list(cmds, count_of(cmds));
-}
-
-void render(uint8_t *buf, struct render_area *area) {
-    // update a portion of the display with a render area
-    uint8_t cmds[] = {
-        SSD1306_SET_COL_ADDR,
-        area->start_col,
-        area->end_col,
-        SSD1306_SET_PAGE_ADDR,
-        area->start_page,
-        area->end_page
-    };
-    
-    SSD1306_send_cmd_list(cmds, count_of(cmds));
-    SSD1306_send_buf(buf, area->buflen);
-}
-
-static void SetPixel(uint8_t *buf, int x,int y, bool on) {
-    assert(x >= 0 && x < SSD1306_WIDTH && y >=0 && y < SSD1306_HEIGHT);
-
-    // The calculation to determine the correct bit to set depends on which address
-    // mode we are in. This code assumes horizontal
-
-    // The video ram on the SSD1306 is split up in to 8 rows, one bit per pixel.
-    // Each row is 128 long by 8 pixels high, each byte vertically arranged, so byte 0 is x=0, y=0->7,
-    // byte 1 is x = 1, y=0->7 etc
-
-    // This code could be optimised, but is like this for clarity. The compiler
-    // should do a half decent job optimising it anyway.
-
-    const int BytesPerRow = SSD1306_WIDTH ; // x pixels, 1bpp, but each row is 8 pixel high, so (x / 8) * 8
-
-    int byte_idx = (y / 8) * BytesPerRow + x;
-    uint8_t byte = buf[byte_idx];
-
-    if (on)
-        byte |=  1 << (y % 8);
-    else
-        byte &= ~(1 << (y % 8));
-
-    buf[byte_idx] = byte;
-}
-
 
 void ssd1306_70_40_i2c_initScreen(void) {
 
     bi_decl(bi_2pins_with_func(22, 23, GPIO_FUNC_I2C));
     bi_decl(bi_program_description("SSD1306 OLED driver I2C example for the Raspberry Pi Pico"));
-
-    printf("Hello, SSD1306 OLED display! Look at my raspberries..\n");
-
     // I2C is "open drain", pull ups to keep signal high when no data is being
     // sent
     i2c_init(i2c1, SSD1306_I2C_CLK * 1000);
@@ -256,48 +49,16 @@ void ssd1306_70_40_i2c_initScreen(void) {
     // run through the complete initialization process
     SSD1306_init();
 
-    // Initialize render area for entire frame (SSD1306_WIDTH pixels by SSD1306_NUM_PAGES pages)
-    struct render_area frame_area = {
-        start_col: 0,
-        end_col : SSD1306_WIDTH - 1,
-        start_page : 0,
-        end_page : SSD1306_NUM_PAGES - 1
-        };
-
     calc_render_area_buflen(&frame_area);
+    ssd1306_70_40_i2c_clearScreen();
+    SSD1306_i2c_render(buf, &frame_area);
 
-    // zero the entire display
-    uint8_t buf[SSD1306_BUF_LEN];
-    memset(buf, 0, SSD1306_BUF_LEN);
-    render(buf, &frame_area);
-
-    // intro sequence: flash the screen 3 times
-    for (int i = 0; i < 3; i++) {
-        SSD1306_send_cmd(SSD1306_SET_ALL_ON);    // Set all pixels on
-        sleep_ms(500);
-        SSD1306_send_cmd(SSD1306_SET_ENTIRE_ON); // go back to following RAM for pixel state
-        sleep_ms(500);
-    }
-    
-    sleep_ms(3000);
-
-    SetPixel(buf, 28,24, true);
-    SetPixel(buf, 40,40, true);
-    SetPixel(buf, 99,63, true);
-    SetPixel(buf, 28,63, true);
-    SetPixel(buf, 99,24, true);
-
-    render(buf, &frame_area);
-
-    // Test the display invert function
-    sleep_ms(3000);
-    SSD1306_send_cmd(SSD1306_SET_INV_DISP);
-    sleep_ms(3000);
-    SSD1306_send_cmd(SSD1306_SET_NORM_DISP);
+    // SSD1306_send_cmd(SSD1306_SET_ALL_OFF);    // Set all pixels on
 }
 
 void ssd1306_70_40_i2c_handleFrameStart(uint8_t frame) {
-    ssd1306_clear(&disp);
+    nearestNeighborHandleFrameStart();
+    ssd1306_70_40_i2c_clearScreen();
 }
 
 void ssd1306_70_40_i2c_blit(uint16_t* downsampled_line, int scanline) {
@@ -308,13 +69,8 @@ void ssd1306_70_40_i2c_blit(uint16_t* downsampled_line, int scanline) {
         
         // comment out if using dithering
         downsampled_pixel = colorToGreyscale(downsampled_pixel);
-
-        
-        if (downsampled_pixel > 127) {
-            ssd1306_draw_pixel(&disp, START_X + x, START_Y + scanline);
-        }
+        SSD1306_setPixel(buf, START_X + x, START_Y + scanline, downsampled_pixel > 127);
     }
-    // ssd1306_show(&disp);
 }
 
 
@@ -323,5 +79,5 @@ void ssd1306_70_40_i2c_handleScanline(uint16_t *line, int scanline) {
 }
 
 void ssd1306_70_40_i2c_handleFrameEnd(uint8_t frame) {
-    ssd1306_show(&disp);
+    SSD1306_i2c_render(buf, &frame_area);
 }
