+++ src/pico/pico-screens/screens/ssd1306_70_40_i2c.h
@@ -8,6 +8,8 @@
 #include "pico.h"
 #include "hardware/i2c.h"
 #include "pico/binary_info.h"
+#include "pico/stdlib.h"
+
 
 #include <stdlib.h>
